Back-End/requirements.txt
Flask==2.2.2
matplotlib==3.5.1
numpy==1.21.5
pandas==1.2.4
PyYAML==6.0
rarfile==4.0
scikit_learn==1.2.1
